-added multiply scripts deleting.

+++ DaemonsRunner.WPF/ViewModels/ScriptsPanelViewModel.cs
@@ -102,6 +102,7 @@ internal partial class ScriptsPanelViewModel : ObservableRecipient
 		{
 			StartScriptsCommand.NotifyCanExecuteChanged();
 			StopScriptsCommand.NotifyCanExecuteChanged();
+			RemoveScriptCommand.NotifyCanExecuteChanged();
 		};
 
 		_dataBus = dataBus;
@@ -123,18 +124,26 @@ internal partial class ScriptsPanelViewModel : ObservableRecipient
 	[RelayCommand(CanExecute = nameof(CanRemoveScript))]
     private async Task RemoveScript()
     {
+		if (RemoveScriptCommand.IsRunning)
+		{
+			return;
+		}
+
         using var scope = _serviceScopeFactory.CreateScope();
         var service = scope.ServiceProvider.GetRequiredService<IScriptService>();
-        var response = await service.DeleteAsync(SelectedScript!.ScriptId);
 
-		if (response.OperationStatus is StatusCode.Success)
+		foreach (var script in Scripts.Where(e => e.IsSelected).ToList())
 		{
-			await App.Current.Dispatcher.InvokeAsync(() => Scripts.Remove(SelectedScript));
+			var response = await service.DeleteAsync(SelectedScript!.ScriptId);
+			if (response.OperationStatus is StatusCode.Success)
+			{
+				await App.Current.Dispatcher.InvokeAsync(() => Scripts.Remove(SelectedScript));
+			}
+			_dataBus.Send(response.Description);
 		}
-		_dataBus.Send(response.Description);
-    }
+	}
 
-    private bool CanRemoveScript() => SelectedScript is not null;
+    private bool CanRemoveScript() => SelectedScript is not null && RunningScripts.Count == 0;
 
     [RelayCommand(CanExecute = nameof(OnCanStartScripts))]
     private async Task StartScripts()
